listo el primero no entiendo porque sale un error con los archivos espero que se pueda arreglar

+++ Producto.hpp
@@ -16,7 +16,7 @@ class Producto {
   
   public:
   Producto();  
-  Producto(int,string,string,int,float);
+  Producto(int,string,string,int,int);
   ~Producto();
   int getID();
   string getNombre();
